+++ packages/expo-splash-screen/android/src/main/java/expo/modules/splashscreen/singletons/SplashScreen.kt
@@ -3,10 +3,7 @@ package expo.modules.splashscreen.singletons
 import android.app.Activity
 import android.util.Log
 import android.view.ViewGroup
-import expo.modules.splashscreen.NativeResourcesBasedSplashScreenViewProvider
-import expo.modules.splashscreen.SplashScreenController
-import expo.modules.splashscreen.SplashScreenImageResizeMode
-import expo.modules.splashscreen.SplashScreenViewProvider
+import expo.modules.splashscreen.*
 import org.unimodules.core.interfaces.SingletonModule
 import java.util.*
 
@@ -17,7 +14,7 @@ object SplashScreen : SingletonModule {
     return "SplashScreen"
   }
 
-  private val controllers = WeakHashMap<Activity, SplashScreenController>()
+  private val controllers = WeakHashMap<Activity, SplashScreenViewController>()
 
   /**
    * Show SplashScreen by mounting it in ContentView.
@@ -42,16 +39,12 @@ object SplashScreen : SingletonModule {
     successCallback: () -> Unit = {},
     failureCallback: (reason: String) -> Unit = { Log.w(TAG, it) }
   ) {
-    // SplashScreen.show can only be called once per activity
-    if (controllers.containsKey(activity)) {
-      return failureCallback("'SplashScreen.show' has already been called for this activity.")
-    }
 
     SplashScreenStatusBar.configureTranslucent(activity, statusBarTranslucent)
 
-    val controller = SplashScreenController(activity, rootViewClass, splashScreenViewProvider)
-    controllers[activity] = controller
-    controller.showSplashScreen(successCallback)
+    val splashView = splashScreenViewProvider.createSplashScreenView(activity)
+    val controller = SplashScreenViewController(activity, rootViewClass, splashView)
+    show(activity, controller, statusBarTranslucent, successCallback, failureCallback)
   }
 
   /**
@@ -80,6 +73,38 @@ object SplashScreen : SingletonModule {
     failureCallback: (reason: String) -> Unit = { Log.w(TAG, it) }
   ) {
     show(activity, splashScreenViewProvider, rootViewClass, statusBarTranslucent, successCallback, failureCallback)
+  }
+
+  /**
+   * Show SplashScreen by mounting it in ContentView.
+   *
+   * Default method for mounting SplashScreen in your app.
+   *
+   * @param activity Target Activity for SplashScreen to be mounted in.
+   * @param SplashScreenViewController SplashScreenViewController to manage the rootView and splashView
+   * @param statusBarTranslucent Flag determining StatusBar translucency in a way ReactNative see it.
+   * @param successCallback Callback to be called once SplashScreen is mounted in view hierarchy.
+   * @param failureCallback Callback to be called once SplashScreen cannot be mounted.
+   * @throws [expo.modules.splashscreen.exceptions.NoContentViewException] when [SplashScreen.show] is called before [Activity.setContentView] (when no ContentView is present for given activity).
+   */
+  @JvmStatic
+  @JvmOverloads
+  fun show(
+  activity: Activity,
+  splashScreenViewController: SplashScreenViewController,
+  statusBarTranslucent: Boolean,
+  successCallback: () -> Unit = {},
+  failureCallback: (reason: String) -> Unit = { Log.w(TAG, it) }
+  ) {
+    // SplashScreen.show can only be called once per activity
+    if (controllers.containsKey(activity)) {
+      return failureCallback("'SplashScreen.show' has already been called for this activity.")
+    }
+
+    SplashScreenStatusBar.configureTranslucent(activity, statusBarTranslucent)
+
+    controllers[activity] = splashScreenViewController
+    splashScreenViewController.showSplashScreen(successCallback)
   }
 
   /**